fix: menu shortcuts and user options

+++ src/App.tsx
@@ -49,26 +49,42 @@ function App() {
 
     if (cookieId) {
       // Update localStorage if cookie ID is present
-      if (localStorageData) {
-        const parsedData = JSON.parse(localStorageData);
-        if (parsedData) {
+      try {
+        const parsedData = localStorageData
+          ? JSON.parse(localStorageData)
+          : null;
+
+        // Only update localStorage if the state has changed
+        if (
+          !parsedData ||
+          JSON.stringify(parsedData) !== JSON.stringify(state)
+        ) {
           localStorage.setItem("yashAppData", JSON.stringify(state));
         }
-      } else {
-        localStorage.setItem("yashAppData", JSON.stringify(state));
+      } catch (error) {
+        console.error("Error parsing localStorage data:", error);
+        localStorage.removeItem("yashAppData"); // Optionally clear corrupt data
       }
     } else {
       // Clear localStorage if cookie ID is not present
       localStorage.clear();
-      // Generate a new unique ID and set it in the cookie
+
+      // Generate a new unique ID and set it in the cookie (expires in 2 days)
       const newId = Date.now().toString();
-      Cookies.set("appId", newId, { expires: 2 / 24 }); // Set cookie for 7 days
+      Cookies.set("appId", newId, { expires: 2 });
+
+      // Set localStorage with the current state
       localStorage.setItem("yashAppData", JSON.stringify(state));
     }
   }, [state]);
 
   useEffect(() => {
-    sessionStorage.setItem("yashAppMsgData", JSON.stringify(messageState));
+    // Compare messageState with sessionStorage to avoid unnecessary updates
+    const existingMsgData = sessionStorage.getItem("yashAppMsgData");
+
+    if (existingMsgData !== JSON.stringify(messageState)) {
+      sessionStorage.setItem("yashAppMsgData", JSON.stringify(messageState));
+    }
   }, [messageState]);
 
   return (
